+++ src/transcendental_factor.cpp
@@ -2238,9 +2238,9 @@ std::vector<std::shared_ptr<SymbolicExpr>> factor_transcendental(
     BigInt target = mignotte * BigInt(2);
     int64_t prime = berl_result.prime;
     int lift_bound = 1;
-    BigInt pk(prime);
+    BigInt pk(static_cast<long long>(prime));
     while (pk <= target && lift_bound < 100) {
-        pk = pk * BigInt(prime);
+        pk = pk * BigInt(static_cast<long long>(prime));
         lift_bound++;
     }
